refactor: Decompose large admin handler module
- Split 1576-line admin/mod.rs into focused modules:
  - common.rs: Shared constants and helper functions
  - users.rs: User management and archival (6 functions)
  - attendance.rs: Admin attendance operations (3 functions)
  - holidays.rs: Holiday management (6 functions)
  - export.rs: Data export functionality (1 function)
- Updated main.rs routing to use new module paths
- Maintained backward compatibility via pub use
- Fixed import issues (double super references)

Issue: #151
