+++ backend/src/handlers/admin/export.rs
@@ -1,0 +1,137 @@
+use axum::{
+    extract::{Extension, Query, State},
+    http::StatusCode,
+    Json,
+};
+use serde_json::{json, Value};
+use sqlx::{PgPool, Postgres, QueryBuilder, Row};
+
+use crate::{
+    config::Config,
+    models::user::User,
+    utils::{csv::append_csv_row, time},
+};
+use super::common::{parse_date_value, push_clause, ExportQuery, bad_request};
+
+/// Export attendance data as CSV
+pub async fn export_data(
+    State((pool, config)): State<(PgPool, Config)>,
+    Extension(user): Extension<User>,
+    Query(q): Query<ExportQuery>,
+) -> Result<Json<Value>, (StatusCode, Json<Value>)> {
+    if !user.is_admin() {
+        return Err((StatusCode::FORBIDDEN, Json(json!({"error":"Forbidden"}))));
+    }
+    
+    // Build filtered SQL
+    let parsed_from = match q.from.as_deref() {
+        Some(raw) => parse_date_value(raw)
+            .ok_or((
+                StatusCode::BAD_REQUEST,
+                Json(json!({"error":"`from` must be a valid date"})),
+            ))
+            .map(Some)?,
+        None => None,
+    };
+    let parsed_to = match q.to.as_deref() {
+        Some(raw) => parse_date_value(raw)
+            .ok_or((
+                StatusCode::BAD_REQUEST,
+                Json(json!({"error":"`to` must be a valid date"})),
+            ))
+            .map(Some)?,
+        None => None,
+    };
+    if let (Some(from), Some(to)) = (parsed_from, parsed_to) {
+        if from > to {
+            return Err((
+                StatusCode::BAD_REQUEST,
+                Json(json!({"error":"`from` must be on or before `to`"})),
+            ));
+        }
+    }
+
+    let mut builder: QueryBuilder<Postgres> = QueryBuilder::new(
+        "SELECT u.username, u.full_name, a.date, a.clock_in_time, a.clock_out_time, a.total_work_hours, a.status \
+         FROM attendance a JOIN users u ON a.user_id = u.id",
+    );
+    let mut has_clause = false;
+    if let Some(ref u_name) = q.username {
+        push_clause(&mut builder, &mut has_clause);
+        builder.push("u.username = ").push_bind(u_name);
+    }
+    if let Some(from) = parsed_from {
+        push_clause(&mut builder, &mut has_clause);
+        builder.push("a.date >= ").push_bind(from);
+    }
+    if let Some(to) = parsed_to {
+        push_clause(&mut builder, &mut has_clause);
+        builder.push("a.date <= ").push_bind(to);
+    }
+    builder.push(" ORDER BY a.date DESC, u.username");
+    let data = builder.build().fetch_all(&pool).await.map_err(|_| {
+        (
+            StatusCode::INTERNAL_SERVER_ERROR,
+            Json(json!({"error": "Database error"})),
+        )
+    })?;
+
+    // Convert to CSV format
+    let mut csv_data = String::new();
+    append_csv_row(
+        &mut csv_data,
+        &[
+            "Username".to_string(),
+            "Full Name".to_string(),
+            "Date".to_string(),
+            "Clock In".to_string(),
+            "Clock Out".to_string(),
+            "Total Hours".to_string(),
+            "Status".to_string(),
+        ],
+    );
+
+    for record in data {
+        let username = record.try_get::<String, _>("username").unwrap_or_default();
+        let full_name = record.try_get::<String, _>("full_name").unwrap_or_default();
+        let date = record.try_get::<String, _>("date").unwrap_or_default();
+        let clock_in = record
+            .try_get::<Option<chrono::NaiveDateTime>, _>("clock_in_time")
+            .ok()
+            .flatten()
+            .map(|t| t.format("%H:%M:%S").to_string())
+            .unwrap_or_default();
+        let clock_out = record
+            .try_get::<Option<chrono::NaiveDateTime>, _>("clock_out_time")
+            .ok()
+            .flatten()
+            .map(|t| t.format("%H:%M:%S").to_string())
+            .unwrap_or_default();
+        let total_hours = record
+            .try_get::<f64, _>("total_work_hours")
+            .map(|h| format!("{:.2}", h))
+            .unwrap_or_else(|_| "0.00".to_string());
+        let status = record.try_get::<String, _>("status").unwrap_or_default();
+
+        append_csv_row(
+            &mut csv_data,
+            &[
+                username,
+                full_name,
+                date,
+                clock_in,
+                clock_out,
+                total_hours,
+                status,
+            ],
+        );
+    }
+
+    Ok(Json(json!({
+        "csv_data": csv_data,
+        "filename": format!(
+            "attendance_export_{}.csv",
+            time::now_in_timezone(&config.time_zone).format("%Y%m%d_%H%M%S")
+        )
+    })))
+}
