+++ documents/feature-expansion-roadmap.md
@@ -1,0 +1,270 @@
+# Timekeeper 機能拡張ロードマップ
+
+## 概要
+
+Timekeeper 勤怠管理システムの機能拡張計画を、重要度と時系列で整理したロードマップです。現在のソリッドなアーキテクチャを基盤に、エンタープライズ利用に向けた段階的機能追加を提案します。
+
+## プロジェクト現状
+
+### スケール metrics
+- **バックエンド**: 65 Rustファイル、13,090行
+- **フロントエンド**: 115 Rustファイル、12,422行  
+- **DBマイグレーション**: 19 SQLファイル
+- **技術スタック**: Axum + SQLx + Leptos/WASM + TailwindCSS
+
+### 既存主要機能
+- 認証: JWT + MFA (TOTP)
+- 勤怠管理: 出勤/退勤打刻、休憩記録
+- 申請管理: 休暇・残業申請と承認
+- 管理機能: ユーザー管理、データエクスポート、監査ログ
+- 祝日管理: Google Calendar連携、カスタム休日
+
+---
+
+##最重要度（緊急対応：0〜3ヶ月）
+
+### 1. セキュリティ強化
+
+#### 目的
+現行の基本的なセキュリティ機能を強化し、エンタープライズレベルの要件を満たす
+
+#### 具体的機能
+- **レート制限**: 認証エンドポイントのAPIコール制限
+- **パスワードポリシー**: 複雑性要件、有効期限、履歴管理
+- **アカウントロックアウト**: ブルートフォース攻撃対策
+- **セルフサービスパスワードリセット**: メール認証によるパスワード変更
+- **同時セッション制限**: 複数デバイスログイン制御
+- **PIIデータ保護**: 個人情報のマスキング・暗号化
+
+#### 実装優先順位
+1. レート制限（middleware層）
+2. パスワードポリシー（user model + validation）
+3. アカウントロックアウト（auth middleware）
+4. パスワードリセット（auth handler + email）
+
+#### 成功指標
+- 認証関連インシデントが90%以上削減
+- セキュリティ監査合格
+- パスワード関連問い合わせが50%削減
+
+### 2. パフォーマンス改善
+
+#### 目的
+現在のパフォーマンスボトルネックを解消し、大規模データ対応を実現
+
+#### 具体的改善
+- **データベース最適化**: インデックス戦略、クエリチューニング
+- **N+1クエリ解消**: 関連データ一括取得
+- **キャッシュレイヤー**: Redisによる頻繁アクセスデータキャッシュ
+- **フロントエンド最適化**: WASM bundle分割、lazy loading
+- **レスポンスタイム改善**: API平均応答時間を50%以下に
+
+#### 実装優先順位
+1. クエリプロファイリングと改善
+2. キャッシュ戦略実装
+3. フロントエンド最適化
+4. パフォーマンス監視導入
+
+### 3. コード品質向上
+
+#### 目的
+技術的負債を解消し、長期保守性を向上
+
+#### 主要改善項目
+- **ハンドラー分割**: `admin/mod.rs`（1000+行）の機能別分割
+- **サービスレイヤー実装**: ビジネスロジックの抽象化
+- **リポジトリパターン完成**: データアクセス層の一貫性
+- **統合テスト拡充**: E2Eテストカバレッジ90%以上
+- **エラーハンドリング統一**: 一貫したエラーレスポンス設計
+
+#### リファクタリング対象
+- `backend/src/handlers/admin/mod.rs`: モジュール分割
+- `frontend/src/api/client.rs`: 認証サービス分離
+- 共通バリデーションフレームワーク導入
+
+---
+
+## 高重要度（中期計画：3〜6ヶ月）
+
+### 4. モバイル対応強化
+
+#### 目的
+フィールドワーカーやリモートワークに対応するモバイル体験を提供
+
+#### 具体的機能
+- **PWA機能**: オフライン対応、キャッシュ戦略
+- **モバイル最適UI**: タッチフレンドリーなインターフェース
+- **プッシュ通知**: 申請・承認・リマインダー通知
+- **クイックアクション**: ワンタップ出勤/退勤
+- **ジオフェンシング**: GPSベースの出勤验证
+
+#### 実装フェーズ
+1. PWA基盤構築（Service Worker, Manifest）
+2. モバイルUIコンポーネント開発
+3. 通知システム実装
+4. オフライン機能拡充
+
+### 5. 通知システム
+
+#### 目的
+業務プロセスの効率化とユーザーエンゲージメント向上
+
+#### 通知チャネル
+- **メール通知**: SMTPサーバー連携、テンプレート管理
+- **インアプリ通知**: リアルタイム通知表示
+- **外部連携**: Slack/Teams Webhook連携
+- **通知設定**: ユーザーごとの通知制御
+
+#### 通知対象イベント
+- 申請提出・承認・却下
+- 勤怠異常（打刻忘れなど）
+- 期限リマインダー
+- システムメンテナンス
+
+### 6. 高度な勤怠機能
+
+#### 目的
+多様な働き方に対応する柔軟な勤怠管理を実現
+
+#### 機能詳細
+- **シフト管理**: スケジュール作成、ローテーションパターン
+- **タイムシート承認**: 多階層承認ワークフロー
+- **残業ルールエンジン**: 法規準拠の残業計算
+- **プロジェクト時間トラッキング**: クライアント別・案件別工数管理
+- **フレックスタイム対応**: コアタイム・変形労働時間
+
+#### 技術要件
+- 新規テーブル: shifts, timesheets, project_allocations
+- ワークフローエンジン: 状態遷移管理
+- 計算エンジン: 複雑な勤怠ルール処理
+
+---
+
+## 中重要度（長期計画：6〜12ヶ月）
+
+### 7. レポート・分析機能
+
+#### 目的
+データ駆動な経営判断を支援する分析機能を提供
+
+#### 主要機能
+- **KPIダッシュボード**: 生産性・欠勤率・残業時間分析
+- **トレンド分析**: 季節変動、長期傾向の可視化
+- **コスト分析**: 人件費・残業コストの詳細分析
+- **カスタムレポート**: ユーザー定義のレポート作成
+- **エクスポート機能**: PDF/Excel形式でのレポート出力
+
+#### 技術要素
+- データウェアハウス設計
+- BIツール連携（Power BI, Tableau）
+- グラフライブラリ導入
+- 非同期レポート生成
+
+### 8. HR統合機能
+
+#### 目的
+HRシステムの中核として完全な人事管理機能を提供
+
+#### 機能範囲
+- **有給休暇管理**: 日数付与・残高計算・繰越処理
+- **組織階層**: 部署・チーム・役職階層管理
+- **バルク操作**: 一括データ更新・承認処理
+- **文書管理**: 添付ファイル・申請書類管理
+- **従業員プロフィール**: スキル・資格・経歴管理
+
+#### データモデル拡張
+- leave_balances, departments, positions, documents
+- 組織図表示機能
+- 権限階層の細分化
+
+### 9. エンタープライズ機能
+
+#### 目的
+大企業向けの要求要件に対応する
+
+#### 主要機能
+- **SSO対応**: SAML/OIDCプロトコル対応
+- **詳細権限管理**: RBACモデルの実装
+- **外部システム連携**: 給与計算・ERPシステム連携
+- **コンプライアンス自動化**: 労働法規準拠チェック
+- **マルチテナント**: 組織単位でのデータ分離
+
+---
+
+## 低重要度（将来検討：12ヶ月〜）
+
+### 10. 高度なコンプライアンス
+
+#### 目的
+グローバル展開に向けたコンプライアンス要件対応
+
+- **GDPR/CCPA完全対応**: データ主体権利の実装
+- **データ匿名化**: テストデータ・分析データの匿名化
+- **地域準拠労働法**: 国別労働法規対応
+- **監査証跡強化**: 変更履歴の詳細追跡
+
+### 11. スケーラビリティ向上
+
+#### 目的
+大規模ユーザー（10,000+）対応
+
+- **マイクロサービス化**: 機能単位のサービス分割
+- **データベースシャーディング**: 水平スケーリング
+- **読み書き分離**: レプリケーションによる負荷分散
+- **バックグラウンド処理**: ジョブキューシステム最適化
+
+---
+
+## 既存課題の対応
+
+### UI/UX改善（既存17件の課題）
+- **空状態コンポーネント**: 視覚的な空状態表示
+- **ボタンホバー効果統一**: 一貫したインタラクション
+- **監査ログUI改善**: フィルタ・ソート・エクスポート機能
+
+### 技術的負債解消
+- **フロントエンドAPIクライアント**: 1400行ファイルの分割
+- **状態管理最適化**: Leptos Signalsの整理
+- **型安全性向上**: serde_json::Valueからの脱却
+
+---
+
+## 実装戦略
+
+### 開発アプローチ
+1. **TDD遵守**: 機能実装前のテスト作成
+2. **段階的リリース**: フィーチャーフラグによる機能切り替え
+3. **並行開発**: バックエンド・フロントエンドの並行実装
+4. **継続的インテグレーション**: 自動テスト・デプロイパイプライン
+
+### リソース配分
+- **Coreチーム**: セキュリティ・パフォーマンス改善（3ヶ月）
+- **Featureチーム**: モバイル・通知機能（6ヶ月）
+- **Architectureチーム**: スケーラビリティ設計（並行）
+
+### 成功指標
+- **技術指標**: レスポンスタイム50%改善、テストカバレッジ90%+
+- **ビジネス指標**: ユーザー満足度80%+、業務効率30%向上
+- **セキュリティ指標**: セキュリティインシデント90%削減
+
+---
+
+## リスク管理
+
+### 技術的リスク
+- **WASM成熟度**: ブラウザ互換性問題
+- **Rustエコシステム**: ライブラリの成熟度不足
+- **データマイグレーション**: 大規模データ移行のリスク
+
+### 緩和策
+- **プルーフオブコンセプト**: 新技術の事前検証
+- **段階的移行**: データ移行のフェーズ化
+- **ロールバック計画**: 任意フェーズでの戻し対応
+
+---
+
+## 結論
+
+Timekeeperの現行アーキテクチャは堅固であり、段階的機能拡張の良好な基盤となります。本ロードマップに従い、重要度に応じた機能拡張を実施することで、中小企業から大企業まで対応する包括的な勤怠管理システムへと発展させることが可能です。
+
+まずはセキュリティ強化とパフォーマンス改善に集中し、その後モバイル対応と高度機能を段階的に追加していくことで、安定性と革新性を両立した製品開発が実現できます。
